requirements.txt
# Discord Issue Dashboard Dependencies
# Generated: 2026-02-04

# Web Framework
streamlit>=1.51.0

# Data Processing
pandas>=2.3.0
numpy>=2.2.0

# Visualization
plotly>=6.3.0

# Google Sheets API
gspread>=6.0.0
google-auth>=2.0.0

# Excel Support (for reference project compatibility)
openpyxl>=3.1.0

# Database
# SQLite is built-in, no extra dependency needed
